chore: extracted name canvas to new component
