+++ src/pages/main/Main.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Canvas, Presentation } from '../../shared/components';
+import { NameCanvas, Presentation } from '../../shared/components';
 
 import styles from './Main.css';
 
@@ -7,7 +7,7 @@ const Main = () => {
 
 	return (
 		<div className={ styles.root }>
-			<Canvas />
+			<NameCanvas />
 			<Presentation />
 		</div>
 	);
