+++ src/shared/components/index.js
@@ -1,2 +1,6 @@
 export { default as Presentation } from './presentation';
 export { default as Canvas } from './canvas';
+export {
+	NameCanvas,
+	ProjectsCanvas,
+} from './canvas/extended';
